requirements.txt
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
filelock==3.18.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
paramiko==3.5.1
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
sshtunnel==0.4.0
urllib3==2.4.0
Werkzeug==3.1.3